guard for magisk user

+++ post-fs-data.sh
@@ -9,11 +9,13 @@ if [ "$(magisk -V)" -lt 26302 ] || [ "$(/data/adb/ksud -V)" -lt 10818 ]; then
 fi
 
 # Remove Play Services from Magisk DenyList when set to Enforce in normal mode
-if magisk --denylist status; then
-    magisk --denylist rm com.google.android.gms
-else # Check if Shamiko is installed and whitelist feature isn't enabled
-    if [ -d "/data/adb/modules/zygisk_shamiko" ] && [ ! -f "/data/adb/shamiko/whitelist" ]; then
-        magisk --denylist add com.google.android.gms com.google.android.gms.unstable
+if command -v magisk >/dev/null; then
+    if magisk --denylist status; then
+        magisk --denylist rm com.google.android.gms
+    else # Check if Shamiko is installed and whitelist feature isn't enabled
+        if [ -d "/data/adb/modules/zygisk_shamiko" ] && [ ! -f "/data/adb/shamiko/whitelist" ]; then
+            magisk --denylist add com.google.android.gms com.google.android.gms.unstable
+        fi
     fi
 fi
